requirements.txt
graph-parser==0.0.7
matplotlib==3.8.4
networkx==3.3
pillow==10.3.0
sympy==1.12
